+++ src/system-agent/operations.test.ts
@@ -795,7 +795,9 @@ describe("parseSystemAgentOperation", () => {
         approved: true,
         deps: { runConfigSet },
       }),
-    ).rejects.toThrow("openclaw onboard");
+      // Denylisted roots cite their documented escalation; route paths point
+      // at the verified set_default_model/onboard flows.
+    ).rejects.toThrow(/openclaw onboard|trusted shell/);
 
     expect(runConfigSet).not.toHaveBeenCalled();
     expect(lines.join("\n")).not.toContain("[openclaw] running:");
